requirements.txt
altgraph==0.16.1
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
Click==7.0
colorama==0.4.3
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.5
djangorestframework==3.11.0
entrypoints==0.3
fbs==0.8.4
future==0.18.2
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab-server==1.0.6
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
pandocfilters==1.4.2
parso==0.5.2
pefile==2019.4.18
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.2
Pygments==2.5.2
PyInstaller==3.5
PyQt5==5.13.2
PyQt5-sip==12.7.0
pyqt5-tools==5.13.2.1.6rc1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==18.1.1
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
